working on report

+++ config/js/calculateHelpers.js
@@ -3,7 +3,7 @@ module.exports = {
     dateFormat: (date) => {
         return new Date(date).toLocaleDateString('en-GB', {
             day: 'numeric',
-            month: 'short',
+            month: 'numeric',
             year: 'numeric',
         })
     },
